+++ Changelog.md
@@ -8,6 +8,7 @@
     - Entry : Add Ctrl-C and Ctrl-V
     - Console : Create console widget
     - MultilineLabel : Create Label widget for multiline
+    - Unit Tests : Fix Vec2 and Color tests
     - Unit Tests : Add Console and MultilineLabel tests
     
     - Optimization